+++ R/http-put.r
@@ -14,12 +14,7 @@
 #' PUT(b2, body = list(x = "A simple text string"), encode = "json")
 PUT <- function(url = NULL, config = list(), ..., body = NULL,
                   encode = c("multipart", "form", "json"),
-                  multipart = TRUE, handle = NULL) {
-  if (!missing(multipart)) {
-    warning("multipart is deprecated, please use encode argument instead",
-      call. = FALSE)
-    encode <- if (multipart) "multipart" else "form"
-  }
+                  handle = NULL) {
   encode <- match.arg(encode)
 
   hu <- handle_url(handle, url, ...)